+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.6.0 (2026-05-07)
+
+- Allow setting long access tokens (when token splits in more than one
+  write syscall).
+- Add Databricks Unified Auth support (Needs root permissions if running multiuser).
+
+
 # 0.5.3 (2026-02-24)
 
 - Improved documentation